Split post-processing and camera
