+++ index.js
@@ -13,6 +13,7 @@ const createAnimation = require('./animation')
 const createGeometry = require('./geometry')
 const createMaterial = require('./material')
 const createCamera = require('./camera')
+const createPostProcessing = require('./post-processing')
 const createOrbiter = require('./orbiter')
 const createAmbientLight = require('./ambient-light')
 const createDirectionalLight = require('./directional-light')
@@ -549,11 +550,12 @@ Renderer.prototype.drawMeshes = function (camera, shadowMapping, shadowMappingLi
   }
 
   if (camera) {
-    if (camera.postprocess && camera.ssao) {
-      sharedUniforms.uAO = camera._frameAOTex
+    const postProcessingCmp = camera.entity.getComponent('PostProcessing')
+    if (postProcessingCmp && postProcessingCmp.ssao) {
+      sharedUniforms.uAO = postProcessingCmp._frameAOTex
       sharedUniforms.uScreenSize = [ camera.viewport[2], camera.viewport[3] ] // TODO: should this be camera viewport size?
     }
-    if (!camera.postprocess) {
+    if (!postProcessingCmp || (postProcessingCmp && !postProcessingCmp.enabled)) {
       sharedUniforms.uExposure = camera.exposure
     }
   }
@@ -697,6 +699,7 @@ Renderer.prototype.drawMeshes = function (camera, shadowMapping, shadowMappingLi
         depthPassOnly: true
       })
     } else {
+      const postProcessingCmp = camera.entity.getComponent('PostProcessing')
       pipeline = this.getGeometryPipeline(geometry, material, skin, {
         numAmbientLights: ambientLights.length,
         numDirectionalLights: directionalLights.length,
@@ -704,8 +707,8 @@ Renderer.prototype.drawMeshes = function (camera, shadowMapping, shadowMappingLi
         numSpotLights: spotLights.length,
         numAreaLights: areaLights.length,
         useReflectionProbes: reflectionProbes.length, // TODO: reflection probes true
-        useSSAO: camera.postprocess && camera.ssao,
-        useTonemapping: !camera.postprocess
+        useSSAO: postProcessingCmp && postProcessingCmp.enabled && postProcessingCmp.ssao,
+        useTonemapping: !postProcessingCmp || (postProcessingCmp && !postProcessingCmp.enabled)
       })
     }
 
@@ -801,9 +804,10 @@ Renderer.prototype.draw = function () {
     const screenSize = [camera.viewport[2], camera.viewport[3]]
     const halfScreenSize = [Math.floor(camera.viewport[2] / 2), Math.floor(camera.viewport[3] / 2)]
     const halfViewport = [0, 0, Math.floor(camera.viewport[2] / 2), Math.floor(camera.viewport[3] / 2)]
-    if (camera.postprocess) {
+    const postProcessingCmp = camera.entity.getComponent('PostProcessing')
+    if (postProcessingCmp && postProcessingCmp.enabled) {
       if (State.profiler) State.profiler.time('depthPrepass', true)
-      ctx.submit(camera._drawFrameNormalsFboCommand, () => {
+      ctx.submit(postProcessingCmp._drawFrameNormalsFboCommand, () => {
         const far = camera.far
         // TODO: Far clipping plane scaling fixes depth buffer precision artifacts
         // but breaks shadows on large scale scenes (eg maps)
@@ -813,9 +817,9 @@ Renderer.prototype.draw = function () {
       })
       if (State.profiler) State.profiler.timeEnd('depthPrepass')
     }
-    if (camera.postprocess && camera.ssao) {
+    if (postProcessingCmp && postProcessingCmp.enabled && postProcessingCmp.ssao) {
       if (State.profiler) State.profiler.time('ssao', true)
-      ctx.submit(camera._ssaoCmd, {
+      ctx.submit(postProcessingCmp._ssaoCmd, {
         uniforms: {
           uNear: camera.near,
           uFar: camera.far,
@@ -824,40 +828,40 @@ Renderer.prototype.draw = function () {
           uInverseViewMatrix: mat4.invert(mat4.copy(camera.viewMatrix)),
           viewProjectionInverseMatrix: mat4.invert(mat4.mult(mat4.copy(camera.viewMatrix), camera.projectionMatrix)), // TODO: GC
           cameraPositionWorldSpace: camera.entity.transform.worldPosition,
-          uIntensity: camera.ssaoIntensity,
+          uIntensity: postProcessingCmp.ssaoIntensity,
           uNoiseScale: [10, 10],
-          uSampleRadiusWS: camera.ssaoRadius,
-          uBias: camera.ssaoBias,
+          uSampleRadiusWS: postProcessingCmp.ssaoRadius,
+          uBias: postProcessingCmp.ssaoBias,
           uScreenSize: screenSize
         }
       })
       if (State.profiler) State.profiler.timeEnd('ssao')
       if (State.profiler) State.profiler.time('ssao-blur', true)
-      ctx.submit(camera._bilateralBlurHCmd, {
+      ctx.submit(postProcessingCmp._bilateralBlurHCmd, {
         uniforms: {
           near: camera.near,
           far: camera.far,
-          sharpness: camera.ssaoBlurSharpness,
+          sharpness: postProcessingCmp.ssaoBlurSharpness,
           imageSize: screenSize,
           depthMapSize: screenSize,
-          direction: [camera.ssaoBlurRadius, 0]
+          direction: [postProcessingCmp.ssaoBlurRadius, 0]
         }
       })
-      ctx.submit(camera._bilateralBlurVCmd, {
+      ctx.submit(postProcessingCmp._bilateralBlurVCmd, {
         uniforms: {
           near: camera.near,
           far: camera.far,
-          sharpness: camera.ssaoBlurSharpness,
+          sharpness: postProcessingCmp.ssaoBlurSharpness,
           imageSize: screenSize,
           depthMapSize: screenSize,
-          direction: [0, camera.ssaoBlurRadius]
+          direction: [0, postProcessingCmp.ssaoBlurRadius]
         }
       })
       if (State.profiler) State.profiler.timeEnd('ssao-blur')
     }
     if (State.profiler) State.profiler.time('drawFrame', true)
-    if (camera.postprocess) {
-      ctx.submit(camera._drawFrameFboCommand, () => {
+    if (postProcessingCmp && postProcessingCmp.enabled) {
+      ctx.submit(postProcessingCmp._drawFrameFboCommand, () => {
         this.drawMeshes(camera, false, null, null, skyboxes[0], false)
       })
     } else {
@@ -867,82 +871,82 @@ Renderer.prototype.draw = function () {
     }
     if (State.profiler) State.profiler.timeEnd('drawFrame')
     if (State.profiler) State.profiler.time('postprocess')
-    if (camera.postprocess && camera.bloom) {
-      ctx.submit(camera._thresholdCmd, {
+    if (postProcessingCmp && postProcessingCmp.enabled && postProcessingCmp.bloom) {
+      ctx.submit(postProcessingCmp._thresholdCmd, {
         uniforms: {
           uExposure: camera.exposure,
-          uBloomThreshold: camera.bloomThreshold,
+          uBloomThreshold: postProcessingCmp.bloomThreshold,
           imageSize: halfScreenSize
         },
         viewport: halfViewport
       })
 
       for (let i = 0; i < 5; i++) {
-        ctx.submit(camera._bloomHCmd, {
+        ctx.submit(postProcessingCmp._bloomHCmd, {
           uniforms: {
-            direction: [camera.bloomRadius, 0],
+            direction: [postProcessingCmp.bloomRadius, 0],
             imageSize: halfScreenSize
           },
           viewport: halfViewport
         })
-        ctx.submit(camera._bloomVCmd, {
+        ctx.submit(postProcessingCmp._bloomVCmd, {
           uniforms: {
-            direction: [0, camera.bloomRadius],
+            direction: [0, postProcessingCmp.bloomRadius],
             imageSize: halfScreenSize
           },
           viewport: halfViewport
         })
       }
     }
-    if (camera.postprocess && camera.dof) {
+    if (postProcessingCmp && postProcessingCmp.enabled && postProcessingCmp.dof) {
       if (State.profiler) State.profiler.time('dof', true)
-      for (let i = 0; i < camera.dofIterations; i++) {
-        ctx.submit(camera._dofBlurHCmd, {
+      for (let i = 0; i < postProcessingCmp.dofIterations; i++) {
+        ctx.submit(postProcessingCmp._dofBlurHCmd, {
           uniforms: {
             near: camera.near,
             far: camera.far,
             sharpness: 0,
             imageSize: screenSize,
             depthMapSize: screenSize,
-            direction: [camera.dofRadius, 0],
-            uDOFDepth: camera.dofDepth,
-            uDOFRange: camera.dofRange
+            direction: [postProcessingCmp.dofRadius, 0],
+            uDOFDepth: postProcessingCmp.dofDepth,
+            uDOFRange: postProcessingCmp.dofRange
           }
         })
-        ctx.submit(camera._dofBlurVCmd, {
+        ctx.submit(postProcessingCmp._dofBlurVCmd, {
           uniforms: {
             near: camera.near,
             far: camera.far,
             sharpness: 0,
             imageSize: screenSize,
             depthMapSize: screenSize,
-            direction: [0, camera.dofRadius],
-            uDOFDepth: camera.dofDepth,
-            uDOFRange: camera.dofRange
+            direction: [0, postProcessingCmp.dofRadius],
+            uDOFDepth: postProcessingCmp.dofDepth,
+            uDOFRange: postProcessingCmp.dofRange
           }
         })
       }
       if (State.profiler) State.profiler.timeEnd('dof')
     }
-    if (camera.postprocess) {
-      ctx.submit(camera._blitCmd, {
+    if (postProcessingCmp && postProcessingCmp.enabled) {
+      ctx.submit(postProcessingCmp._blitCmd, {
         uniforms: {
-          uExposure: camera.exposure,
-          uFXAA: camera.fxaa,
-          uFog: camera.fog,
-          uBloom: camera.bloom,
-          uBloomIntensity: camera.bloomIntensity,
           uNear: camera.near,
           uFar: camera.far,
           uFov: camera.fov,
-          uSunDispertion: camera.sunDispertion,
-          uSunIntensity: camera.sunIntensity,
-          uSunColor: camera.sunColor,
-          uInscatteringCoeffs: camera.inscatteringCoeffs,
-          uFogColor: camera.fogColor,
-          uFogStart: camera.fogStart,
-          uFogDensity: camera.fogDensity,
-          uSunPosition: camera.sunPosition,
+          uExposure: camera.exposure,
+          uFXAA: postProcessingCmp.fxaa,
+          uFog: postProcessingCmp.fog,
+          uBloom: postProcessingCmp.bloom,
+          uBloomIntensity: postProcessingCmp.bloomIntensity,
+          uSunDispertion: postProcessingCmp.sunDispertion,
+          uSunIntensity: postProcessingCmp.sunIntensity,
+          uSunColor: postProcessingCmp.sunColor,
+          uInscatteringCoeffs: postProcessingCmp.inscatteringCoeffs,
+          uFogColor: postProcessingCmp.fogColor,
+          uFogStart: postProcessingCmp.fogStart,
+          uFogDensity: postProcessingCmp.fogDensity,
+          uSunPosition: postProcessingCmp.sunPosition,
           uOutputEncoding: ctx.Encoding.Gamma,
           uScreenSize: screenSize
         },
@@ -1017,6 +1021,10 @@ Renderer.prototype.material = function (opts) {
 
 Renderer.prototype.camera = function (opts) {
   return createCamera(Object.assign({ ctx: this._ctx, rgbm: State.rgbm }, opts))
+}
+
+Renderer.prototype.postProcessing = function (opts) {
+  return createPostProcessing(Object.assign({ ctx: this._ctx, rgbm: State.rgbm }, opts))
 }
 
 Renderer.prototype.orbiter = function (opts) {
